client/requirements.txt
pyqt6
websockets
pyyaml
numpy
pyaudio
